LayoutTests/platform/ios-sim/fast/forms/textarea-input-event-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=15189

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL handleTextareaOnInputCallCount should be 1. Was 0.
FAIL handleDivOnInputCallCount should be 1. Was 0.
FAIL handleTextareaOnInputCallCount should be 1. Was 0.
FAIL handleDivOnInputCallCount should be 1. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
